chore: Improve several aspects of the m3u8 module
Enhancements have been made to the m3u8 module. The target_duration function now accepts a u64 value instead of a u32, expanding possible duration values. We've also added `#[allow(clippy::too_many_arguments)]` to bypass Clippy's warnings for the media and stream_inf function. Also, more meaningful error messages have been added and the code is now streamlined for improved tag extraction. For validation.rs, we replaced the u32 data type with the u64 in InvalidTargetDuration to match the change in builder.rs.

+++ src/m3u8/playlist/builder.rs
@@ -78,7 +78,7 @@ impl PlaylistBuilder {
     }
 
     /// Adds an `ExtXTargetDuration` tag.
-    pub fn target_duration(mut self, duration: u32) -> Self {
+    pub fn target_duration(mut self, duration: u64) -> Self {
         self.tags.push(Tag::ExtXTargetDuration(duration));
         self
     }
@@ -185,6 +185,7 @@ impl PlaylistBuilder {
     }
 
     /// Adds an `ExtXMedia` tag.
+    #[allow(clippy::too_many_arguments)]
     pub fn media(
         mut self,
         type_: String,
@@ -210,6 +211,7 @@ impl PlaylistBuilder {
     }
 
     /// Adds an `ExtXStreamInf` tag.
+    #[allow(clippy::too_many_arguments)]
     pub fn stream_inf(
         mut self,
         bandwidth: u32,
